+++ changelog.md
@@ -6,8 +6,8 @@ compromise uses semver, and pushes to npm and github frequently
 
 While all _Major_ releases should be reviewed, our only _large_ releases are **v6** in 2016 **v12** in 2019 and **v14** in 2022. Others have been mostly incremental.
 
-<!-- #### 14.4.0 [Unreleased]
-
+<!-- #### 14.5.0 [Unreleased]
+- **[change]** - improvements to negative-optional match logic - `!foo?`
 -->
 
 #### 14.4.0 [July 2021]